NOJIRA: Refactor Payload client tests
